work!

+++ entrypoint.sh
@@ -10,15 +10,16 @@ curl -sqL "https://steamcdn-a.akamaihd.net/client/installer/steamcmd_linux.tar.g
 shopt -s nullglob # allow pattern matching failure to omit result instead of returning the pattern used
 
 # Update server
-rm -rf /home/container/steamapps
 
-if [[ "${VALIDATE_SERVER_FILES}" == "1" ]]; then
-	./steamcmd/steamcmd.sh +force_install_dir /home/container +login anonymous +app_update 258550 validate +quit
-else
-	./steamcmd/steamcmd.sh +force_install_dir /home/container +login anonymous +app_update 258550 +quit
+if [[ "${STEAMCMD_UPDATE_SERVER}" == "1"]]; then
+	rm -rf /home/container/steamapps
+
+	if [[ "${VALIDATE_SERVER_FILES}" == "1" ]]; then
+		./steamcmd/steamcmd.sh +force_install_dir /home/container +login anonymous +app_update 258550 validate +quit
+	else
+		./steamcmd/steamcmd.sh +force_install_dir /home/container +login anonymous +app_update 258550 +quit
+	fi
 fi
-
-MODIFIED_STARTUP=$(eval echo "${STARTUP}" | sed -e 's/{{/${/g' -e 's/}}/}/g') # replaces {{var}} to ${var} (so bash can understand it)
 
 # Update carbon (uses carbon production)
 curl -sSL "https://github.com/CarbonCommunity/Carbon/releases/download/production_build/Carbon.Linux.Release.tar.gz" | tar zx
@@ -27,11 +28,15 @@ curl -sSL "https://github.com/CarbonCommunity/Carbon/releases/download/productio
 mkdir -p HarmonyMods
 cp /LinuxStdinSupport.dll ./HarmonyMods/LinuxStdinSupport.dll
 
-# cp /start_server.sh ./start_server.sh
 
-# Run server
-chmod +x ./carbon.sh
+# Hijack carbon's startup command so we can inject our own script
 sed -i "10c\
 source ./start_server.sh
 " carbon.sh
+if [[ "${USING_ORIGINAL_START_SERVER_SCRIPT}" == "1"]]; then
+	cp /start_server.sh ./start_server.sh
+fi
+
+# Run server
+chmod +x ./carbon.sh
 ./carbon.sh
